chore: move instrumentations into a new module
