+++ packages/opentelemetry/test/unit/lib/index.spec.js
@@ -2,7 +2,6 @@ jest.mock('../../../package.json', () => ({
 	name: 'mock-package',
 	version: '1.2.3'
 }));
-jest.mock('@opentelemetry/auto-instrumentations-node');
 jest.mock('@opentelemetry/exporter-trace-otlp-proto');
 jest.mock('@opentelemetry/exporter-trace-otlp-proto/package.json', () => ({
 	name: 'mock-otel-tracing-package',
@@ -13,20 +12,21 @@ jest.mock('@opentelemetry/sdk-node');
 jest.mock('@dotcom-reliability-kit/app-info');
 jest.mock('@opentelemetry/api');
 jest.mock('@dotcom-reliability-kit/logger');
-jest.mock('@dotcom-reliability-kit/log-error');
-jest.mock('@dotcom-reliability-kit/errors');
+jest.mock('../../../lib/config/instrumentations', () => ({
+	createInstrumentationConfig: jest
+		.fn()
+		.mockReturnValue('mock-instrumentations')
+}));
 jest.mock('../../../lib/config/resource', () => ({
 	createResourceConfig: jest.fn().mockReturnValue('mock-resource')
 }));
 
+const {
+	createInstrumentationConfig
+} = require('../../../lib/config/instrumentations');
 const { createResourceConfig } = require('../../../lib/config/resource');
 const { diag, DiagLogLevel } = require('@opentelemetry/api');
-const {
-	getNodeAutoInstrumentations
-} = require('@opentelemetry/auto-instrumentations-node');
 const logger = require('@dotcom-reliability-kit/logger');
-const { logRecoverableError } = require('@dotcom-reliability-kit/log-error');
-const { UserInputError } = require('@dotcom-reliability-kit/errors');
 const opentelemetrySDK = require('@opentelemetry/sdk-node');
 const appInfo = require('@dotcom-reliability-kit/app-info');
 const {
@@ -46,7 +46,6 @@ appInfo.environment = 'MOCK_ENVIRONMENT';
 
 logger.createChildLogger.mockReturnValue('mock child logger');
 DiagLogLevel.INFO = 'mock info log level';
-getNodeAutoInstrumentations.mockReturnValue('mock implementation response');
 
 // Import the OTel function for testing
 const openTelemetryTracing = require('../../../lib/index');
@@ -79,27 +78,20 @@ describe('@dotcom-reliability-kit/opentelemetry', () => {
 			expect(opentelemetrySDK.NodeSDK.prototype.start).toHaveBeenCalledTimes(1);
 		});
 
+		it('configures the OpenTelemetry instrumentations', () => {
+			expect(createInstrumentationConfig).toHaveBeenCalledTimes(1);
+			expect(createInstrumentationConfig).toHaveBeenCalledWith();
+			expect(
+				opentelemetrySDK.NodeSDK.mock.calls[0][0].instrumentations
+			).toStrictEqual('mock-instrumentations');
+		});
+
 		it('configures the OpenTelemetry resource', () => {
 			expect(createResourceConfig).toHaveBeenCalledTimes(1);
 			expect(createResourceConfig).toHaveBeenCalledWith();
 			expect(opentelemetrySDK.NodeSDK.mock.calls[0][0].resource).toStrictEqual(
 				'mock-resource'
 			);
-		});
-
-		it('sets OpenTelemetry instrumentations by auto-instrumenting common and built-in Node.js modules', () => {
-			expect(getNodeAutoInstrumentations).toHaveBeenCalledTimes(1);
-			expect(getNodeAutoInstrumentations).toHaveBeenCalledWith({
-				'@opentelemetry/instrumentation-http': {
-					ignoreIncomingRequestHook: expect.any(Function)
-				},
-				'@opentelemetry/instrumentation-fs': {
-					enabled: false
-				}
-			});
-			expect(
-				opentelemetrySDK.NodeSDK.mock.calls[0][0].instrumentations
-			).toEqual(expect.arrayContaining(['mock implementation response']));
 		});
 
 		it('creates traces via an instantiated OTLPTraceExporter', () => {
@@ -137,87 +129,6 @@ describe('@dotcom-reliability-kit/opentelemetry', () => {
 
 		it('it does not create traces via an instantiated NoopSpanProcessor', () => {
 			expect(NoopSpanProcessor).toHaveBeenCalledTimes(0);
-		});
-
-		describe('ignoreIncomingRequestHook filter', () => {
-			let ignoreIncomingRequestHook;
-
-			beforeAll(() => {
-				ignoreIncomingRequestHook =
-					getNodeAutoInstrumentations.mock.calls[0][0][
-						'@opentelemetry/instrumentation-http'
-					].ignoreIncomingRequestHook;
-			});
-
-			it('returns `true` with a request to `/__gtg`', () => {
-				const mockRequest = {
-					url: '/__gtg?a=b',
-					headers: { host: 'mock-host' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(true);
-			});
-
-			it('returns `true` with a request to `/__health`', () => {
-				const mockRequest = {
-					url: '/__health?a=b',
-					headers: { host: 'mock-host' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(true);
-			});
-
-			it('returns `true` with a request to `/favicon.ico`', () => {
-				const mockRequest = {
-					url: '/favicon.ico?a=b',
-					headers: { host: 'mock-host' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(true);
-			});
-
-			it('returns `false` with a request to anything else', () => {
-				const mockRequest = {
-					url: '/mock-endpoint',
-					headers: { host: 'mock-host' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(false);
-			});
-
-			it('returns `false` when a request URL is not present', () => {
-				const mockRequest = {
-					url: undefined,
-					headers: { host: 'mock-host' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(false);
-			});
-
-			it("doesn't throw when the host header isn't set", () => {
-				const mockRequest = {
-					url: '/mock-endpoint',
-					headers: {}
-				};
-				expect(() => ignoreIncomingRequestHook(mockRequest)).not.toThrow();
-			});
-
-			it('logs a warning and returns `false` when a request URL cannot be parsed', () => {
-				const mockRequest = {
-					url: '/mock-endpont',
-					headers: { host: '' }
-				};
-				expect(ignoreIncomingRequestHook(mockRequest)).toBe(false);
-				expect(UserInputError).toHaveBeenCalledTimes(1);
-				expect(UserInputError).toHaveBeenCalledWith(
-					expect.objectContaining({
-						code: 'OTEL_REQUEST_FILTER_FAILURE'
-					})
-				);
-				expect(logRecoverableError).toHaveBeenCalledTimes(1);
-				expect(logRecoverableError).toHaveBeenCalledWith(
-					expect.objectContaining({
-						error: expect.any(UserInputError),
-						includeHeaders: ['host'],
-						request: mockRequest
-					})
-				);
-			});
 		});
 	});
